vibe-UI-main/backend/user-management-service/requirements.txt
# Dependencies for user-management-service
flask==2.3.2
flask-cors==4.0.0
pymongo==4.4.1
requests==2.31.0
python-dotenv==1.0.0
pyjwt==2.7.0